+++ Source/Modules/Catalog/IO/Gapotchenko.FX.IO/StreamPolyfills.Read.cs
@@ -7,6 +7,7 @@
 // Year of introduction: 2025
 
 using System.Buffers;
+using System.Runtime.InteropServices;
 
 namespace Gapotchenko.FX.IO;
 
@@ -27,12 +28,12 @@ partial class StreamPolyfills
     /// <returns>
     /// The total number of bytes read into the buffer.
     /// This can be less than the size of the buffer if that many bytes are not currently available,
-    /// or zero (<c>0</c>) if the buffer's length is zero or the end of the stream has been reached.
+    /// or <c>0</c> (zero) if the buffer's length is <c>0</c> or the end of the stream has been reached.
     /// </returns>
 #if TFF_STREAM_SPAN
     [EditorBrowsable(EditorBrowsableState.Never)]
 #else
-    // TODO: method use should be discouraged because the polyfill implementation incurs memory allocations and copying.
+    // TODO: method use should be discouraged because the polyfill implementation incurs memory copying.
 #endif
     public static int Read(
 #if !TFF_STREAM_SPAN
@@ -48,19 +49,79 @@ partial class StreamPolyfills
 #else
         int count = buffer.Length;
 
-        var pool = ArrayPool<byte>.Shared;
-        byte[] array = pool.Rent(count);
+        var arrayPool = ArrayPool<byte>.Shared;
+        byte[] array = arrayPool.Rent(count);
         try
         {
-            int bytesRead = stream.Read(array, 0, count);
-            if (bytesRead > 0)
-                array.AsSpan(0, bytesRead).CopyTo(buffer);
-            return bytesRead;
+            int result = stream.Read(array, 0, count);
+            new ReadOnlySpan<byte>(array, 0, result).CopyTo(buffer);
+            return result;
         }
         finally
         {
-            pool.Return(array);
+            arrayPool.Return(array);
         }
 #endif
     }
+
+#if TFF_VALUETASK
+
+    /// <summary>
+    /// Asynchronously reads a sequence of bytes from the current stream,
+    /// advances the position within the stream by the number of bytes read,
+    /// and monitors cancellation requests.
+    /// </summary>
+    /// <param name="stream">The stream to read the bytes from.</param>
+    /// <param name="buffer">The region of memory to write the data into.</param>
+    /// <param name="cancellationToken">The cancellation token.</param>
+    /// <returns>
+    /// A task that represents the asynchronous read operation.
+    /// The value of its <see cref="ValueTask{TResult}.Result"/> property contains the total number of bytes read into the buffer.
+    /// The result value can be less than the length of the buffer if that many bytes are not currently available,
+    /// or it can be <c>0</c> (zero) if the length of the buffer is <c>0</c> or if the end of the stream has been reached.
+    /// </returns>
+#if TFF_STREAM_SPAN
+    [EditorBrowsable(EditorBrowsableState.Never)]
+#else
+    // TODO: method use should be discouraged because the polyfill implementation incurs memory copying.
+#endif
+    public static ValueTask<int> ReadAsync(
+#if !TFF_STREAM_SPAN
+        this
+#endif
+        Stream stream,
+        Memory<byte> buffer,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(stream);
+
+#if TFF_STREAM_SPAN
+        return stream.ReadAsync(buffer, cancellationToken);
+#else
+        if (MemoryMarshal.TryGetArray(buffer, out ArraySegment<byte> segment))
+            return new(stream.ReadAsync(segment.Array!, segment.Offset, segment.Count, cancellationToken));
+        else
+            return CoreImpl(stream, buffer, cancellationToken);
+
+        static async ValueTask<int> CoreImpl(Stream stream, Memory<byte> buffer, CancellationToken cancellationToken)
+        {
+            int count = buffer.Length;
+
+            var arrayPool = ArrayPool<byte>.Shared;
+            byte[] array = arrayPool.Rent(count);
+            try
+            {
+                int result = await stream.ReadAsync(array, 0, count, cancellationToken).ConfigureAwait(false);
+                new ReadOnlySpan<byte>(array, 0, result).CopyTo(buffer.Span);
+                return result;
+            }
+            finally
+            {
+                arrayPool.Return(array);
+            }
+        }
+#endif
+    }
+
+#endif
 }